requirements.txt
streamlit==1.29.0
pandas==2.1.1
requests==2.31.0
streamlit-option-menu==0.3.6
streamlit-folium==0.15.0
matplotlib==3.8.0
folium==0.14.0
Pillow==10.0.1
numpy==1.26.0
plotly==5.17.0
scikit-learn==1.3.0
qrcode==7.4.2
statsmodels==0.14.0
prophet==1.1.4
xgboost==1.7.6
networkx==3.1
scipy==1.11.3
openpyxl==3.1.2
twilio==8.3.0
python-dotenv==1.0.0
